Add files via upload
So everything we have worked on is within the "newWebsite" file. There might be some things we need to change like how the images are displayed- i read on the google doc file that you wanted us to use a link for the images but here we just used the direct source.